确认才会执行 分开两个命令
